+++ js/components/ad-form.js
@@ -1,5 +1,4 @@
 import {dropValidity} from '../utils/helpers.js';
-import * as Forms from '../common/forms.js';
 import {TOKIO_LAT, TOKIO_LNG} from '../constants/constants.js';
 
 const MIN_PRICE = {
@@ -18,6 +17,7 @@ const price = adForm.querySelector('#price');
 const timeIn = adForm.querySelector('#timein');
 const timeOut = adForm.querySelector('#timeout');
 const address = adForm.querySelector('#address');
+const interactiveControls = adForm.querySelectorAll('input, select, fieldset');
 
 const _validateField = (field, checkupFunction) => {
   const customValidityMessage = checkupFunction();
@@ -77,7 +77,7 @@ const validateAdForm = (evt) => {
 };
 
 const setAddress = (lat, lng) => {
-  address.value = `lat: ${lat}, lng: ${lng}`;
+  address.value = `${lat.toFixed(5)}, ${lng.toFixed(5)}`;
 };
 
 const activate = () => {
@@ -96,8 +96,8 @@ const activate = () => {
 
   adForm.addEventListener('submit', validateAdForm);
 
-  Forms.activate(adForm, 'ad-form--disabled');
-};
+  adForm.classList.remove('ad-form--disabled');
+  interactiveControls.forEach((el) => {el.removeAttribute('disabled', '');});};
 
 const deactivate = () => {
   hostType.removeEventListener('input', setupMinPrice);
@@ -111,7 +111,8 @@ const deactivate = () => {
   capacity.removeEventListener('input', validateCapacity);
   roomNumber.removeEventListener('input', validateRoomNumber);
 
-  Forms.deactivate(adForm, 'ad-form--disabled');
+  adForm.classList.add('ad-form--disabled');
+  interactiveControls.forEach((el) => {el.setAttribute('disabled', '');});
 };
 
 export {activate, deactivate, setAddress};
